Moved calculator to form, removed everything other than radiobuttons, added block comment explaining radiobuttons
