+++ src/components/forms/radioButtonGroup/components/RadioButton.tsx
@@ -1,5 +1,6 @@
-import React from "react";
-import styles from "./radioButton.module.css";
+import React from 'react';
+import styles from 'src/components/forms/radioButtonGroup/radioButtonGroup.module.css';
+import textStyles from "src/components/text/text.module.css";
 
 interface RadioButtonProps {
   id: string;
@@ -8,7 +9,7 @@ interface RadioButtonProps {
   value?: string;
   checked?: boolean;
   disabled?: boolean;
-  defaultChecked?: boolean; 
+  defaultChecked?: boolean;
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
 }
 
@@ -23,24 +24,26 @@ export const RadioButton: React.FC<RadioButtonProps> = ({
   onChange,
 }) => {
   return (
-    <div
-      className={`${styles.main}`}
-    >   
-    <input
-        className={`${styles.input}`}
+    <> 
+      <label
+        htmlFor={id}
+        className={`${styles.container} ${textStyles.caption} ${disabled ? styles.disabledLabel : styles.label}`}
+      >
+      <input
+        className={styles.input}
         type="radio"
         name={name}
         id={id}
         value={value}
         checked={checked}
         disabled={disabled}
-        aria-disabled={disabled ? "true" : "false"}
+        aria-disabled={disabled ? 'true' : 'false'}
         defaultChecked={defaultChecked}
         onChange={onChange}
-      /> 
-      <label htmlFor={id} className={`${styles.label} ${disabled? styles.disabledLabel : styles.label}`}>
-       {label}
+      />
+     
+        {label}
       </label>
-    </div >
+    </>
   );
 };
