+++ src/salaryAndBenefits/SalaryAndBenefits.tsx
@@ -1,0 +1,11 @@
+
+const SalaryAndBenefits = () => {
+   
+    return (
+        <p>
+           Salary and Benefits form  
+        </p>
+    );
+};
+
+export default SalaryAndBenefits;
